requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
langchain==0.2.5
langchain-google-genai==0.0.18
chromadb==0.5.3
numpy==1.26.4
tiktoken==0.7.0
